refactor(workflows): extract MediaWorkflowName and MediaInput into FFmpeg-free types package
Move MediaWorkflowName and MediaInput into workflows/media/types so that
cmd/watcher can import only that package, breaking the transitive CGO
dependency chain through workflows/steps → pkg/ffmpeg → go-astiav.

The watcher binary now builds with CGO_ENABLED=0 and the watcher OCI image
no longer requires libav shared libraries in its closure.
